+++ src/main/scala/dxWDL/RunnerEval.scala
@@ -134,13 +134,13 @@ object RunnerEval {
               jobOutputPath : Path,
               jobInfoPath: Path) : Unit = {
         // Figure out input/output types
-        val (inputTypes, outputTypes) = Utils.loadExecInfo(Utils.readFileContent(jobInfoPath))
+        //val (inputTypes, outputTypes) = Utils.loadExecInfo(Utils.readFileContent(jobInfoPath))
+        val (inputSpec, outputSpec) = Utils.loadExecInfo
 
         // Parse the inputs, do not download files from the platform,
         // they will be passed as links.
         val inputLines : String = Utils.readFileContent(jobInputPath)
-        val inputs: Map[String, WdlVarLinks] = WdlVarLinks.loadJobInputsAsLinks(inputLines,
-                                                                                inputTypes)
+        val inputs: Map[String, WdlVarLinks] = WdlVarLinks.loadJobInputsAsLinks(inputLines, inputSpec)
         appletLog(s"Initial inputs=${inputs}")
 
         // make sure the workflow elements are all declarations
@@ -152,7 +152,7 @@ object RunnerEval {
         val outputs : Map[String, BValue] = evalDeclarations(decls, inputs, false, None)
 
         // Keep only exported variables
-        val exported = outputs.filter{ case (varName, _) => outputTypes contains varName }
+        val exported = outputs.filter{ case (varName, _) => outputSpec contains varName }
         val outputFields: Map[String, JsonNode] = exported.map {
             case (varName, bValue) => WdlVarLinks.genFields(bValue.wvl, varName)
         }.flatten.toMap