+++ RELEASE_NOTES.md
@@ -2,6 +2,8 @@
 
 ## 0.44
 - Use hashes instead of files for non-native dx types
+- Do not use the help field in applet input/output arguments to carry
+  WDL typing information.
 
 ## 0.43
 - Speed up the creation of a dx:workflow. With one API call create